requirements.txt
boto3==1.35.57
cohere==5.13.3
matplotlib==3.7.1
numpy==1.26.0
opencv-python==4.10.0.84
paddleocr==2.9.1
pandas==2.2.3
Pillow==10.4.0
streamlit==1.40.2
